Fix #28: Add a procedure to store jobs in a file when shutting down, so that reminders etc. are not lost (#39)
* Add admin decorator
* WIP
* wip
* Refactor database into a separate file
* Working version
* Add verif alias command
* Add decorator everywhere
* Remove ChatModule because lazy, cf. #22
* bump black
* Add default
* Add migration; Add default in decorator
* Add Cooldown
* Balancing
* Spaaam
* Small fixes
* wip
* Fix #28: Add a procedure to store jobs in a file when shutting down, so that reminders etc. are not lost
* A bit more flexible
