+++ main.py
@@ -11,7 +11,7 @@ import logging
 from telegram.ext import Updater
 
 
-from databases import ChatCommand, User
+from databases import ChatCommand, ChatJob, start_jobs_in_database, User
 from modules.admin import Admin
 from modules.bot import Bot
 from modules.community.exp import Exp
@@ -43,9 +43,9 @@ def main() -> None:
 
     # Commands
     Bot(logger).add_commands(dispatcher)
-    RemindMe(logger).add_commands(dispatcher)
+    RemindMe(logger, table=ChatJob, dispatcher=dispatcher).add_commands(dispatcher)
     Special(logger).add_commands(dispatcher)
-    Admin(logger, table=ChatCommand).add_dispatcher(dispatcher).add_commands(dispatcher)
+    Admin(logger, table=ChatCommand, dispatcher=dispatcher).add_commands(dispatcher)
 
     # Community commands
     Exp(logger, table=User).add_commands(dispatcher)
