+++ modules/community/remindme.py
@@ -10,65 +10,11 @@ from telegram.ext import CallbackContext, CommandHandler
 import dateparser
 
 
+from databases import start_jobs_in_database
+
+
 from ..base import Base, command_enabled
-
-
-def naturaltime(delta):
-    """
-    Return a nice phrasing for the remaining time.
-    :param delta: datetime.timedelta
-    :return: Phrasing: String.
-    """
-    time_strings = {
-        "now": "now",
-        "second": ("in a second", "in {} seconds"),
-        "minute": ("in a minute", "in {} minutes"),
-        "hour": ("in an hour", "in {} hours"),
-        "day": ("in a day", "in {} days"),
-        "week": ("in a week", "in {} weeks"),
-        "month": ("in a month", "in {} months"),
-        "year": ("in a year", "in {} years"),
-    }
-
-    if delta.days != 0:
-        if delta.days < 7:
-            if delta.days == 1:
-                return time_strings["day"][0]
-            else:
-                return time_strings["day"][1].format(delta.days)
-        elif delta.days // 7 < 4:
-            if delta.days // 7 == 1:
-                return time_strings["week"][0]
-            else:
-                return time_strings["week"][1].format(delta.days // 7)
-        elif delta.days // 7 // 4 < 12:
-            if delta.days // 7 // 4 == 1:
-                return time_strings["month"][0]
-            else:
-                return time_strings["month"][1].format(delta.days // 7 // 4)
-        else:
-            if delta.days // 7 // 4 // 12 == 1:
-                return time_strings["year"][0]
-            else:
-                return time_strings["year"][1].format(delta.days // 7 // 4 // 12)
-    else:
-        if delta.seconds == 0:
-            return time_strings["now"]
-        elif delta.seconds < 60:
-            if delta.seconds == 1:
-                return time_strings["second"][0]
-            else:
-                return time_strings["second"][1].format(delta.seconds)
-        elif delta.seconds // 60 < 60:
-            if delta.seconds // 60 == 1:
-                return time_strings["minute"][0]
-            else:
-                return time_strings["minute"][1].format(delta.seconds // 60)
-        else:
-            if delta.seconds // 60 // 60 == 1:
-                return time_strings["hour"][0]
-            else:
-                return time_strings["hour"][1].format(delta.seconds // 60 // 60)
+from .helpers import alarm, naturaltime
 
 
 class RemindMe(Base):
@@ -76,7 +22,7 @@ class RemindMe(Base):
     Reminders and alarms within a groupchat.
     """
 
-    def __init__(self, logger=None):
+    def __init__(self, logger=None, table=None, dispatcher=None):
         commandhandlers = [
             CommandHandler(["remindme", "remind_me", "set", "alarm"], self.remindme),
             CommandHandler(
@@ -88,18 +34,13 @@ class RemindMe(Base):
                 self.allremindme,
             ),
         ]
-        super().__init__(logger, commandhandlers)
+        super().__init__(
+            logger=logger, commandhandlers=commandhandlers, table=table, dispatcher=dispatcher
+        )
+
+        start_jobs_in_database(self.dispatcher, alarm)
 
     # Borrowed from https://github.com/django/django/blob/main/django/contrib/humanize/templatetags/humanize.py#L169, but worse
-
-    def alarm(self, context: CallbackContext) -> None:
-        """
-        Helper function to send the actual alarm message.
-        """
-        job = context.job
-        context.bot.send_message(
-            job.context["chat_id"], text="Reminder!", reply_to_message_id=job.context["message_id"]
-        )
 
     @command_enabled(default=True)
     def remindme(self, update: Update, context: CallbackContext) -> None:
@@ -111,7 +52,6 @@ class RemindMe(Base):
 
         try:
             _, message = update.message.text.split(" ", 1)
-            print(message)
             interpreted = dateparser.parse(message)
             if not interpreted:
                 update.message.reply_text("I didn't understand, sorry...")
@@ -126,19 +66,24 @@ class RemindMe(Base):
                 return
 
             context.job_queue.run_once(
-                self.alarm,
+                alarm,
                 delta.total_seconds(),
                 context={"chat_id": chat_id, "message_id": message_id},
                 name=f"{chat_id}_{message_id}",
+            )
+
+            # Add job to database in case of crash
+            self.table.create(
+                chatid=chat_id, messageid=message_id, deadline=interpreted, fun=alarm.__name__
             )
 
             update.message.reply_text(
                 f"I will remind you this {naturaltime(delta)} ({str(delta).split('.', maxsplit=1)[0]})!"
             )
 
-            self.logger.info(f"{update.effective_user.first_name} set an alarm!")
+            self.logger.info(f"{update.effective_user.first_name} sets an alarm!")
         except (IndexError, ValueError):
-            update.message.reply_text("It seems like you used that command wrong. (:.")
+            update.message.reply_text("It seems like you used that command wrong.")
 
     @command_enabled(default=True)
     def allremindme(self, update: Update, context: CallbackContext) -> None:
